--- plugins/avatar/src/systems/index.ts
@@ -1,1 +1,0 @@
-export { AvatarMainSystem } from './AvatarMainSystem';
