+++ plugins/avatar/src/ui/LocalCursor.tsx
@@ -1,5 +1,5 @@
-import { cursorZIndex, lerpX, lerpY, localAvatar, localCursorStyle, localStatus, scrollX, scrollY } from '../state';
-import { cssToState } from '../systems/AvatarMainSystem';
+import { cursorZIndex, lerpX, lerpY, localAvatar, localCursorStyle, scrollX, scrollY } from '../state';
+import { cssToState } from '../systems/utils';
 
 export const LocalCursor = () => {
     const cursorState = cssToState(localCursorStyle);
@@ -19,25 +19,6 @@ export const LocalCursor = () => {
             }}
         >
             <img src={stateDef.url} alt="cursor" style={{ maxWidth: '64px', maxHeight: '64px', display: 'block' }} />
-            {localStatus === 'busy' && (
-                <div
-                    style={{
-                        position: 'absolute',
-                        top: '-8px',
-                        right: '-8px',
-                        width: '16px',
-                        height: '16px',
-                        borderRadius: '50%',
-                        backgroundColor: '#fa5252',
-                        display: 'flex',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        fontSize: '10px',
-                    }}
-                >
-                    🔴
-                </div>
-            )}
         </div>
     );
 };
