--- plugins/avatar/src/worker.tsx
@@ -1,7 +1,0 @@
-import { resetState } from './state';
-import { AvatarMainSystem } from './systems/AvatarMainSystem';
-
-resetState();
-Ubi.registerSystem(AvatarMainSystem);
-
-console.log('[Avatar Worker] Initialized. Singleton presence mode active.');
